refactor: 메인 페이지 컴포넌트 경로 수정
