+++ src/app/(routes)/landing/page.tsx
@@ -1,6 +1,6 @@
-import LandingHeader from './LandingHeader';
-import LandingFeatures from './LandingFeatures';
-import LandingFooter from './LandingFooter';
+import LandingHeader from '@/app/components/landing/LandingHeader';
+import LandingFeatures from '@/app/components/landing/LandingFeatures';
+import LandingFooter from '@/app/components/landing/LandingFooter';
 
 export default function LandingPage() {
   return (
