+++ src/app/components/landing/LandingFooter.tsx
